+++ buession-core/src/main/java/com/buession/core/deserializer/FastJsonJsonDeserializer.java
@@ -19,7 +19,7 @@
  * +-------------------------------------------------------------------------------------------------------+
  * | License: http://www.apache.org/licenses/LICENSE-2.0.txt 										       |
  * | Author: Yong.Teng <webmaster@buession.com> 													       |
- * | Copyright @ 2013-2024 Buession.com Inc.														       |
+ * | Copyright @ 2013-2025 Buession.com Inc.														       |
  * +-------------------------------------------------------------------------------------------------------+
  */
 package com.buession.core.deserializer;
@@ -39,7 +39,7 @@ public class FastJsonJsonDeserializer extends AbstractJsonDeserializer<JSON> {
 	@Override
 	public <V> V deserialize(final String str) throws DeserializerException {
 		Assert.isNull(str, "String cloud not be null.");
-		return JSON.parseObject(str, new com.alibaba.fastjson.TypeReference<V>() {
+		return JSON.parseObject(str, new com.alibaba.fastjson.TypeReference<>() {
 
 		});
 	}